requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.6
Django==4.1.4
dnspython==2.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
etils==0.8.0
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==1.12
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.11.0
google-api-python-client==2.90.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-functions==1.9.0
google-pasta==0.2.0
googleapis-common-protos==1.57.0
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.51.1
gspread==5.10.0
gspread-dataframe==3.3.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
httplib2==0.22.0
huggingface-hub==0.12.1
idna==3.4
importlib-resources==5.10.0
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langchain==0.0.228
langchainplus-sdk==0.0.20
libclang==14.0.6
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
networkx==2.8.8
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==1.3.4
pandas-stubs==1.2.0.35
parso==0.8.3
passlib==1.7.4
peewee==3.15.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///private/tmp/pillow-20231009-7955-1ba2b4g/Pillow-10.0.1
pipenv==2023.6.2
platformdirs==3.5.1
promise==2.3
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==4.23.2
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11 @ file:///private/tmp/pybind11-20231006-5320-1wy2z7l/pybind11-2.11.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.8
Pygments==2.14.0
pymongo==4.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyvis==0.3.1
PyYAML @ file:///private/tmp/pyyaml-20231006-5597-bdjg8x/PyYAML-6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
seaborn==0.12.2
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.15
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.2.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.7.0
tensorflow-estimator==2.9.0
tensorflow-metal==0.5.0
termcolor==2.0.1
toml==0.10.2
tqdm==4.65.0
traitlets==5.8.1
trio==0.22.2
trio-websocket==0.11.1
types-pytz==2022.6.0.1
typing-inspect==0.9.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.9.0
zope.event==4.6
zope.interface==6.0
